Update SurveyBot.py
Updated the survey under activate_survey(server_name, digit_code) method. It is not for Red Lobster. I am leaving this section filled, so that you can see how the code is to be used. If you want to edit it for your own change the activate_survey snippets.
Also, I added Exit nodes for where we are coming out of through Tor's Network. Added a new interact with method for select combo boxes and text. Finally, I added a flask snippet here so it can be ran on a private local server, connected to ngrok then finally interacted with through a text message via twilio api.